bring in some basic GB and NES libraries
The "Continue" screen is a conventional test ROM. Once the game
is in place, it'll be used to determine how many levels to skip.
It'll use my proportional font code, so bring that in.

The logo (not counting the spinning coin dotting the i) uses 33
unique tiles, 32 not blank.  I may need to find a way to exclude
specific tiles in the conversion.