spec for Game Boy CPU exerciser
There's a lot of red at
<http://tasvideos.org/EmulatorResources/GBAccuracyTests.html>
It looks like a few of the first ten will be CPU bugs, in particular
from the half carry.  Yet blargg's tests are based on the same CRC
principle as ZEXALL, making bad cases hard to find.  So I'll write
a program that lets me key in machine code and run it, varying the
input values and displaying output values in real time.