Add Markov chain surname generator
A suggestion by Basxto in gbdev Discord prompted me to write a
program to search for feminine names and surnames meeting both
the anagram and aesthetic, based on the statistics of British
and American surnames.  This is the first time in my life that
I've implemented the A* search algorithm.

Not sure whether I want her name to be Mindy Beageonton (which
spells Nintendo Game Boy) or Emily F. MacProut (which spells
Family Computer).