add diagram of GB #23 STAT blocking
Lycoder, developer of the Game Boy emulator Geebly, was confused
as to what was going on in GB coin #23 "STAT blocking".  I added an
ASCII art diagram of STAT rising and falling edges to its comments.

On a related note, I've become aware of hardware platforms derived
from Game Boy with changes intended to induce incompatibility with
Game Boy software, presumably to deter one particular litigious
company from suing.  Some of these shuffle the order of registers
in $FF00-$FF7F and bits in LCDC and STAT and/or change the expected
logo data in $0104-$0133.

So along with the diagram, I'm using more symbolic constants rather
than magic numbers for STAT values.  I'm also updating the vendored
hardware.inc to a newer version that adds the needed IEB_* and
STATB_* values.  RGBASM 0.5 or later is required.