turn off SGB flag; describe SGB plans
Turn off the "SGB aware" flag in the header (and coin 28) for now
and describe in gb_coins.md what I seek to test later.

BGB developer beware said on 2022-04-05 and 04-23 that it'd be more
convenient to leave Super Game Boy tests out of Numism.  I decided
to compromise by developing the SGB tests separately to merge later
as forthcoming stage 5.