reorder GBC coins in file; update results
New policy for no$gmb is to treat it as a DMG/SGB emulator.

All emulators get the SGBC coin.  I kept it to see which emulators
have some sort of SGBC mode, as opposed to the less inauthentic
"GBC with initial SGB border" mode found in things like Goomba.