add crash handler
This was used to track down a wayward program counter in SameBoy
when pressing a button for the first time on the Continue menu.
This in turn turned out to be related to random jumps to $0000
at the end of the joypad interrupt handler.

TODO: figure out why it's jumping to $0000 using debugger command
b/j $0000