extractcels.py: pack frames for streaming
A streaming cel is prefixed with the tile numbers it uses, and then
tile numbers in that cel's strips are remapped to indices into that
array (0 to N-1).  This corresponds to copying tiles 0 to N-1
to VRAM and then drawing the metasprite from the copied tiles.

Streaming tile sheets can have 256 tiles.  This is half the size of
Donny from Haunted: Halloween '86 (unless we switch to 8x16-pixel
tiles), and it's a lot smaller than Monster in My Pocket player
characters.  We'll see if that's enough as I box the rest of Mindy.