continue: add heartbeat to cursor
calc84maniac reported that the menu was unresponsive in GEM, a
Game Boy emulator for MSX turbo R published by The New Image.
<http://www.tni.nl/products/gem.html>

It scored 1/10 coins on the first stage.  The coin at top wasn't
spinning (expected with <5 coins), and the cursor wasn't moving.
I realized there was no way to tell whether the menu was actually
frozen or just failing to receive input through continue_read_pad.
Moving the cursor left and right on a 64-pixel cycle looks like
an unobtrusive way to show it continuing to run.